+++ src/js/templates/signup.js
